pulling the main script out of utils/evaluation_manager.py and including it as a CI script. Work needs to be done on MPIEvaluationManagerCI to make it CI ready
