+++ src/hiopbbpy/utils/evaluation_manager.py
@@ -34,196 +34,126 @@ else:
 
 
 class EvaluationManager:
-    """Class that manages the evaluation of functions using multiple executors.
+  """Class that manages the evaluation of functions using multiple executors.
 
-    It allows for the submission of tasks to different executors:
+  It allows for the submission of tasks to different executors:
 
-    - Intra-node parallelism using ProcessPoolExecutor (default)
-    - Inter-node parallelism using MPIPoolExecutor (if available)
+  - Intra-node parallelism using ProcessPoolExecutor (default)
+  - Inter-node parallelism using MPIPoolExecutor (if available)
 
-    Tasks are submitted to the executors, and their results can be retrieved
-    once they are completed. The submission of tasks is asynchronous, and the
-    retrieval of results is done in a blocking manner. The syncronization of all
-    tasks is done using the `sync` method, which waits for all tasks to complete
-    before proceeding.
+  Tasks are submitted to the executors, and their results can be retrieved
+  once they are completed. The submission of tasks is asynchronous, and the
+  retrieval of results is done in a blocking manner. The syncronization of all
+  tasks is done using the `sync` method, which waits for all tasks to complete
+  before proceeding.
 
-    :param cpu_executor: The executor to use for intra-node parallelism, i.e.,
-        that controls tasks on the executor's local node. If None, a new
-        ProcessPoolExecutor is created.
-    :param mpi_executor: The executor to use for inter-node parallelism, i.e.,
-        that controls tasks on the executor's remote nodes. If None, a new
-        MPIPoolExecutor is created. If MPI is not available, this parameter is
-        ignored.
+  :param cpu_executor: The executor to use for intra-node parallelism, i.e.,
+      that controls tasks on the executor's local node. If None, a new
+      ProcessPoolExecutor is created.
+  :param mpi_executor: The executor to use for inter-node parallelism, i.e.,
+      that controls tasks on the executor's remote nodes. If None, a new
+      MPIPoolExecutor is created. If MPI is not available, this parameter is
+      ignored.
 
-    .. attribute:: logger:
+  .. attribute:: logger:
 
-        The logger for the EvaluationManager. It is used to log messages.
+      The logger for the EvaluationManager. It is used to log messages.
 
-    .. attribute:: executors:
-        A dictionary of executors used for task submission. The keys are
-        "cpu" and "mpi", and the values are the corresponding executors.
-        The default is a ProcessPoolExecutor for "cpu" and an MPIPoolExecutor
-        for "mpi" if MPI is available.
+  .. attribute:: executors:
+      A dictionary of executors used for task submission. The keys are
+      "cpu" and "mpi", and the values are the corresponding executors.
+      The default is a ProcessPoolExecutor for "cpu" and an MPIPoolExecutor
+      for "mpi" if MPI is available.
+  """
+
+  def __init__(
+    self,
+    cpu_executor=None,
+    mpi_executor=None) -> None:
+    self._queue = deque([])
+    self.logger = logging.getLogger(self.__class__.__name__)
+
+    self.executors = {
+        "cpu": ProcessPoolExecutor() if cpu_executor is None else cpu_executor
+    }
+    if _EVALUATION_MANAGER_USES_MPI4PY:
+      self.executors["mpi"] = (
+          MPIPoolExecutor() if mpi_executor is None else mpi_executor
+      )
+    elif mpi_executor is not None:
+      self.executors["mpi"] = mpi_executor
+
+    self.logger.info("EvaluationManager initialized with executors:")
+    for key, executor in self.executors.items():
+      self.logger.info(f"  - {key}: {executor}")
+
+  def __del__(self) -> None:
+    for executor in self.executors.values():
+      executor.shutdown(wait=False)
+    self.logger.info("EvaluationManager destroyed and executors shut down.")
+
+  def sync(self) -> None:
+    """Synchronizes all tasks by waiting for their completion.
+
+    Results can be retrieved using :meth:`retrieve_results()`.
     """
+    future_objs = [queue_obj[1] for queue_obj in self._queue]
+    wait(future_objs)
 
-    def __init__(
-        self,
-        cpu_executor=None,
-        mpi_executor=None,
-    ) -> None:
-        self._queue = deque([])
-        self.logger = logging.getLogger(self.__class__.__name__)
+  def submit_tasks(self, fn, X, execute_at="cpu", *args, **kwargs) -> None:
+    """Submits tasks to the specified executor.
 
-        self.executors = {
-            "cpu": ProcessPoolExecutor() if cpu_executor is None else cpu_executor
-        }
-        if _EVALUATION_MANAGER_USES_MPI4PY:
-            self.executors["mpi"] = (
-                MPIPoolExecutor() if mpi_executor is None else mpi_executor
-            )
-        elif mpi_executor is not None:
-            self.executors["mpi"] = mpi_executor
+    :param fn: The function to be executed.
+    :param X: The sequence of input data for the function.
+    :param execute_at: The executor to use for task submission. It can be
+        "cpu" for intra-node parallelism or "mpi" for inter-node
+        parallelism.
+    :param args: Additional positional arguments to be passed to the
+        function.
+    :param kwargs: Additional keyword arguments to be passed to the
+        function.
+    """
+    key = execute_at.lower()
+    for x in X:
+      future_obj = self.executors[key].submit(fn, x, *args, **kwargs)
+      self._queue.append([copy.deepcopy(x), future_obj])
+      self.logger.info(f"Submitted f({x})")
 
-        self.logger.info("EvaluationManager initialized with executors:")
-        for key, executor in self.executors.items():
-            self.logger.info(f"  - {key}: {executor}")
+  def retrieve_results(self) -> tuple[list, list]:
+    """Retrieves the results of completed tasks.
 
-    def __del__(self) -> None:
-        for executor in self.executors.values():
-            executor.shutdown(wait=False)
-        self.logger.info("EvaluationManager destroyed and executors shut down.")
+    :return: A tuple containing two lists: the inputs and the results of
+        the completed tasks.
+    """
+    X = deque([])
+    F = deque([])
+    Idxs = deque([])
+    new_queue = deque([])
+    for i, item in enumerate(self._queue):
+      x = item[0]
+      future = item[1]
+      if future.done():
+        # Try to get result
+        try:
+          fx = future.result()
+          Idxs.append(i)
+        except CancelledError:
+          self.logger.warning(f"The execution of x={x} was cancelled.")
+          continue
 
-    def sync(self) -> None:
-        """Synchronizes all tasks by waiting for their completion.
+        # Add result to the output
+        X.append(x)
+        F.append(fx)
+        self.logger.info(f"Completed: f({x}) = {fx}")
+      else:
+        # Keep the future in the queue
+        new_queue.append(item)
 
-        Results can be retrieved using :meth:`retrieve_results()`.
-        """
-        future_objs = [queue_obj[1] for queue_obj in self._queue]
-        wait(future_objs)
+    # Remove completed tasks from the queue
+    self._queue = new_queue
 
-    def submit_tasks(self, fn, X, execute_at="cpu", *args, **kwargs) -> None:
-        """Submits tasks to the specified executor.
-
-        :param fn: The function to be executed.
-        :param X: The sequence of input data for the function.
-        :param execute_at: The executor to use for task submission. It can be
-            "cpu" for intra-node parallelism or "mpi" for inter-node
-            parallelism.
-        :param args: Additional positional arguments to be passed to the
-            function.
-        :param kwargs: Additional keyword arguments to be passed to the
-            function.
-        """
-        key = execute_at.lower()
-        for x in X:
-            future_obj = self.executors[key].submit(fn, x, *args, **kwargs)
-            self._queue.append([copy.deepcopy(x), future_obj])
-            self.logger.info(f"Submitted f({x})")
-
-    def retrieve_results(self) -> tuple[list, list]:
-        """Retrieves the results of completed tasks.
-
-        :return: A tuple containing two lists: the inputs and the results of
-            the completed tasks.
-        """
-        X = deque([])
-        F = deque([])
-        Idxs = deque([])
-        new_queue = deque([])
-        for i, item in enumerate(self._queue):
-            x = item[0]
-            future = item[1]
-            if future.done():
-                # Try to get result
-                try:
-                    fx = future.result()
-                    Idxs.append(i)
-                except CancelledError:
-                    self.logger.warning(f"The execution of x={x} was cancelled.")
-                    continue
-
-                # Add result to the output
-                X.append(x)
-                F.append(fx)
-                self.logger.info(f"Completed: f({x}) = {fx}")
-            else:
-                # Keep the future in the queue
-                new_queue.append(item)
-
-        # Remove completed tasks from the queue
-        self._queue = new_queue
-
-        X = [X[Idxs[i]] for i in range(len(Idxs))]
-        F = [F[Idxs[i]] for i in range(len(Idxs))]
-        return list(X), list(F)
+    X = [X[Idxs[i]] for i in range(len(Idxs))]
+    F = [F[Idxs[i]] for i in range(len(Idxs))]
+    return list(X), list(F)
 
 
-def _fn_for_test(x, sleep_time=0.1):
-    import time
-
-    if is_running_with_mpi():
-        from mpi4py import MPI
-
-        comm = MPI.COMM_WORLD
-        rank = comm.Get_rank()
-        print(f"Rank {rank}: Processing {x}")
-
-    time.sleep(sleep_time)  # Simulate some work
-    return x * x
-
-
-if __name__ == "__main__":
-    import argparse
-    import time
-
-    # Arguments for command line
-    parser = argparse.ArgumentParser(
-        description="Execute n function calls with t duration.",
-        epilog="To properly run the example with mpi4py, use: env MPI4PY_FUTURES_MAX_WORKERS=<N> mpiexec -n 1 python evaluation_manager.py",
-    )
-    parser.add_argument("-n", type=int, default=100, help="Number of tasks to execute")
-    parser.add_argument(
-        "-t", "--sleep_time", type=float, default=1, help="Sleep time for each task"
-    )
-    args = parser.parse_args()
-
-    if is_running_with_mpi():
-        executor_type = "mpi"
-    else:
-        executor_type = "cpu"
-
-    # Set up logging
-    logging.basicConfig(level=logging.INFO)
-
-    # Create manager
-    manager = EvaluationManager()
-
-    # Submit tasks to the manager
-    t0 = time.perf_counter()
-    manager.submit_tasks(
-        _fn_for_test,
-        [i for i in range(args.n)],
-        execute_at=executor_type,
-        sleep_time=args.sleep_time,
-    )
-
-    # Do some other work while tasks are running
-    print("Doing other work", end="", flush=True)
-    for i in range(5):
-        print(".", end="", flush=True)
-        time.sleep(args.sleep_time)
-    print(" Done.")
-
-    # Wait for all tasks to complete
-    print("Waiting for tasks to complete...")
-    manager.sync()
-    t1 = time.perf_counter()
-
-    # Retrieve and show results
-    X, F = manager.retrieve_results()
-    print("X:", X)
-    print("F:", F)
-    print(f"Total time: {t1 - t0:.2f} seconds")
-
-    # Clean up
-    del manager
